# Splitter throughput improvements