--- a/repo/object/object_writer.go
+++ b/repo/object/object_writer.go
@@ -102,20 +102,36 @@ func (w *objectWriter) Close() error {
 	return nil
 }
 
+func nextSplitPoint(data []byte, s splitter.Splitter) int {
+	for i, b := range data {
+		if s.ShouldSplit(b) {
+			return i + 1
+		}
+	}
+
+	return -1
+}
+
 func (w *objectWriter) Write(data []byte) (n int, err error) {
 	dataLen := len(data)
 	w.totalLength += int64(dataLen)
 
-	for _, d := range data {
-		if err := w.buffer.WriteByte(d); err != nil {
+	for len(data) > 0 {
+		n := nextSplitPoint(data, w.splitter)
+		if n < 0 {
+			// no split points in the buffer
+			w.buffer.Write(data)
+			break
+		}
+
+		// found a split point after `n` bytes, write first n bytes then flush and repeat with the remainder.
+		w.buffer.Write(data[0:n])
+
+		if err := w.flushBuffer(); err != nil {
 			return 0, err
 		}
 
-		if w.splitter.ShouldSplit(d) {
-			if err := w.flushBuffer(); err != nil {
-				return 0, err
-			}
-		}
+		data = data[n:]
 	}
 
 	return dataLen, nil